// dashboard/src/api/models/priceModelEnum.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `flat` - Flat
 * `graduated` - Graduated
 * `volume` - Volume
 */
export type PriceModelEnum =
  (typeof PriceModelEnum)[keyof typeof PriceModelEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PriceModelEnum = {
  flat: "flat",
  graduated: "graduated",
  volume: "volume",
} as const;