// dashboard/src/api/models/productsListStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

export type ProductsListStatus =
  (typeof ProductsListStatus)[keyof typeof ProductsListStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductsListStatus = {
  active: "active",
  archived: "archived",
} as const;